+++ src/auto_smart_factory/src/agent/path_planning/ThetaStarGridNodeInformation.cpp
@@ -1,6 +1,6 @@
 #include "agent/path_planning/ThetaStarGridNodeInformation.h"
 
-ThetaStarGridNodeInformation::ThetaStarGridNodeInformation(const GridNode* node, ThetaStarGridNodeInformation* prev, float time) :
+ThetaStarGridNodeInformation::ThetaStarGridNodeInformation(const GridNode* node, ThetaStarGridNodeInformation* prev, double time) :
 		node(node),
 		time(time),
 		prev(prev),
